requirements.txt
cppimport==22.8.2
joblib==1.3.2
logger==1.4
matplotlib==3.7.2
networkx==3.1
numpy==1.24.3
pandas==2.0.3
PyYAML==6.0.1
PyYAML==6.0.2
scikit_learn==1.3.0
scipy==1.14.0
torch==2.0.1+cu118
torch_geometric==2.3.1
torch_scatter==2.1.1+pt20cu118
torch_sparse==0.6.17+pt20cu118
